--- custom_addons/product_external_id/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_product_external_id
