+++ custom_addons/visible_product_external_id/__init__.py
