+++ custom_addons/visible_product_external_id/models/__init__.py
