+++ custom_addons/visible_product_external_id/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_visible_product_external_id
