+++ custom_addons/visible_product_external_id/views/product_views.xml
@@ -5,8 +5,8 @@
     <!-- ======================= -->
     
     <!-- Product template form view -->
-    <record id="product_template_form_view_inherit_external_id" model="ir.ui.view">
-        <field name="name">product.template.form.inherit.external.id</field>
+    <record id="product_template_form_view_inherit_visible_external_id" model="ir.ui.view">
+        <field name="name">product.template.form.inherit.visible.external.id</field>
         <field name="model">product.template</field>
         <field name="inherit_id" ref="product.product_template_form_view"/>
         <field name="arch" type="xml">
@@ -18,8 +18,8 @@
     </record>
 
     <!-- Product template tree/list view -->
-    <record id="product_template_tree_view_inherit_external_id" model="ir.ui.view">
-        <field name="name">product.template.tree.inherit.external.id</field>
+    <record id="product_template_tree_view_inherit_visible_external_id" model="ir.ui.view">
+        <field name="name">product.template.tree.inherit.visible.external.id</field>
         <field name="model">product.template</field>
         <field name="inherit_id" ref="product.product_template_tree_view"/>
         <field name="arch" type="xml">
@@ -30,8 +30,8 @@
     </record>
 
     <!-- Product template search view -->
-    <record id="product_template_search_view_inherit_external_id" model="ir.ui.view">
-        <field name="name">product.template.search.inherit.external.id</field>
+    <record id="product_template_search_view_inherit_visible_external_id" model="ir.ui.view">
+        <field name="name">product.template.search.inherit.visible.external.id</field>
         <field name="model">product.template</field>
         <field name="inherit_id" ref="product.product_template_search_view"/>
         <field name="arch" type="xml">
@@ -46,8 +46,8 @@
     <!-- ======================= -->
     
     <!-- Product variant form view -->
-    <record id="product_product_form_view_inherit_external_id" model="ir.ui.view">
-        <field name="name">product.product.form.inherit.external.id</field>
+    <record id="product_product_form_view_inherit_visible_external_id" model="ir.ui.view">
+        <field name="name">product.product.form.inherit.visible.external.id</field>
         <field name="model">product.product</field>
         <field name="inherit_id" ref="product.product_normal_form_view"/>
         <field name="arch" type="xml">
@@ -59,8 +59,8 @@
     </record>
 
     <!-- Product variant tree/list view -->
-    <record id="product_product_tree_view_inherit_external_id" model="ir.ui.view">
-        <field name="name">product.product.tree.inherit.external.id</field>
+    <record id="product_product_tree_view_inherit_visible_external_id" model="ir.ui.view">
+        <field name="name">product.product.tree.inherit.visible.external.id</field>
         <field name="model">product.product</field>
         <field name="inherit_id" ref="product.product_product_tree_view"/>
         <field name="arch" type="xml">
@@ -71,8 +71,8 @@
     </record>
 
     <!-- Product variant search view -->
-    <record id="product_product_search_view_inherit_external_id" model="ir.ui.view">
-        <field name="name">product.product.search.inherit.external.id</field>
+    <record id="product_product_search_view_inherit_visible_external_id" model="ir.ui.view">
+        <field name="name">product.product.search.inherit.visible.external.id</field>
         <field name="model">product.product</field>
         <field name="inherit_id" ref="product.product_search_form_view"/>
         <field name="arch" type="xml">
